--- src/services/moderation/adminReportModeration.service.ts
@@ -1,486 +1,0 @@
-import HttpError from "../../utils/httpError.util.js";
-import { clearReportsCache } from "../../utils/clearCache.util.js";
-import { makeJobId } from "../../utils/makeJobId.util.js";
-
-import Report from "../../models/report.model.js";
-
-import prisma from "../../config/prisma.config.js";
-
-import publishSocketDisconnect from "../../utils/publishSocketDisconnect.util.js";
-
-import moderationMetricsQueue from "../../queues/moderationMetrics.queue.js";
-import moderationAuditQueue from "../../queues/moderationAudit.queue.js";
-import deleteContentQueue from "../../queues/deleteContent.queue.js";
-
-import applyAiModerationDecisionService from "./applyAiModerationDecision.service.js";
-import routeNotification from "../notification/routeNotification.service.js";
-
-import crypto from "crypto";
-
-type AdminReportActionTaken = "BAN_TEMP" | "BAN_PERM" | "WARN" | "IGNORE";
-
-type ReportTargetType = "QUESTION" | "ANSWER" | "REPLY" | "AI_ANSWER_FEEDBACK";
-
-const actionToModerationStatus: Record<
-  AdminReportActionTaken,
-  "APPROVED" | "FLAGGED" | "REJECTED"
-> = {
-  BAN_TEMP: "REJECTED",
-  BAN_PERM: "REJECTED",
-  WARN: "FLAGGED",
-  IGNORE: "APPROVED",
-};
-
-const adminModerateReport = async ({
-  targetId,
-  targetType,
-  reviewedBy,
-  reviewComment,
-  actionTaken,
-  title,
-  reasons,
-  banDurationMs,
-  warningDurationMs,
-}: {
-  targetId: string;
-  targetType: ReportTargetType;
-  reviewedBy: string;
-  reviewComment?: string;
-  actionTaken: AdminReportActionTaken;
-  title: string;
-  reasons?: string[];
-  banDurationMs?: number;
-  warningDurationMs?: number;
-}) => {
-  if (actionTaken === "BAN_TEMP") {
-    if (banDurationMs === undefined) {
-      throw new HttpError("banDurationMs is required for BAN_TEMP", 400);
-    }
-  }
-
-  if (actionTaken === "WARN") {
-    if (warningDurationMs === undefined) {
-      throw new HttpError("warningDurationMs is required for WARN", 400);
-    }
-  }
-
-  const foundReport = await Report.findOne({
-    _id: targetId,
-    targetType,
-    status: "PENDING",
-  });
-
-  if (!foundReport) throw new HttpError("Report not found", 404);
-
-  if (foundReport.targetUserId === reviewedBy) {
-    throw new HttpError("Self-moderation is not allowed", 403);
-  }
-
-  const targetUser = await prisma.user.findUnique({
-    where: { id: foundReport.targetUserId as string },
-    select: { status: true },
-  });
-
-  if (!targetUser) throw new HttpError("Target user not found", 404);
-
-  if (actionTaken !== "IGNORE" && targetUser.status === "TERMINATED") {
-    throw new HttpError("Target user account is already terminated", 409);
-  }
-
-  const resolvedAt = new Date();
-  const reportTargetUserId = foundReport.targetUserId as string;
-  const reportContentId = String(foundReport.targetId);
-  const shouldRemoveContent =
-    actionTaken === "BAN_PERM" || actionTaken === "BAN_TEMP";
-
-  const decisionId = crypto.randomUUID();
-  const reportId = String(foundReport.id);
-  const reporterUserId = foundReport.reportedBy as string;
-
-  const claimReport = await Report.findOneAndUpdate(
-    {
-      _id: foundReport._id,
-      status: "PENDING",
-      reviewedBy: null,
-    },
-    {
-      reviewedBy,
-      reviewComment,
-      reviewedAt: resolvedAt,
-    },
-    { new: true },
-  );
-
-  if (!claimReport) {
-    throw new HttpError("Report already resolved", 409);
-  }
-
-  const updateReportStatus = async (
-    status: "RESOLVED" | "DISMISSED",
-    actionTaken: AdminReportActionTaken,
-    meta: any,
-  ) => {
-    const updatedReport = await Report.findOneAndUpdate(
-      { _id: foundReport._id, status: "PENDING", reviewedBy },
-      {
-        status,
-        actionTaken,
-        isRemovingContent: shouldRemoveContent,
-      },
-      { new: true },
-    );
-
-    if (!updatedReport) throw new HttpError("Report already resolved", 409);
-
-    await moderationAuditQueue.add(
-      "UPDATE_REPORT_STATUS",
-      {
-        decisionId,
-        targetType: "REPORT",
-        targetId: updatedReport.id,
-        targetUserId: updatedReport.targetUserId,
-        actorType: "ADMIN_MODERATION",
-        adminId: reviewedBy,
-        actionTaken,
-        meta,
-      },
-      {
-        removeOnComplete: true,
-        removeOnFail: false,
-        jobId: makeJobId("moderationAudit", decisionId, "updateReportStatus"),
-      },
-    );
-
-    await routeNotification({
-      recipientId: reporterUserId,
-      actorId: reviewedBy,
-      event: "REPORT_UPDATE",
-      target: {
-        entityType: "REPORT",
-        entityId: reportId,
-      },
-      meta: {
-        reportId,
-
-        status,
-        actionTaken,
-        isRemovingContent: updatedReport.isRemovingContent,
-
-        targetContentId: reportContentId,
-        targetContentType: targetType,
-      },
-    });
-  };
-
-  const queueDeleteContentIfNeeded = async (meta: any) => {
-    if (!shouldRemoveContent) return;
-
-    await deleteContentQueue.add(
-      "REMOVE_MODERATED_CONTENT",
-      {
-        userId: reportTargetUserId,
-        targetType,
-        targetId: reportContentId,
-      },
-      {
-        removeOnComplete: true,
-        removeOnFail: false,
-        jobId: makeJobId(
-          "deleteContent",
-          decisionId,
-          "REMOVE_MODERATED_CONTENT",
-          targetType,
-          reportContentId,
-        ),
-      },
-    );
-
-    await moderationAuditQueue.add(
-      "REMOVE_CONTENT",
-      {
-        decisionId,
-        targetType: "CONTENT",
-        targetId: foundReport.targetId,
-        targetUserId: foundReport.targetUserId,
-        actorType: "ADMIN_MODERATION",
-        adminId: reviewedBy,
-        actionTaken: "REMOVE",
-        meta,
-      },
-      {
-        removeOnComplete: true,
-        removeOnFail: false,
-        jobId: makeJobId("moderationAudit", decisionId, "removeContent"),
-      },
-    );
-
-    await routeNotification({
-      recipientId: reportTargetUserId,
-      actorId: reviewedBy,
-      event: "REMOVE_CONTENT",
-      target: {
-        entityType: targetType,
-        entityId: reportContentId,
-      },
-      meta: {
-        reportId,
-        targetType,
-        actionTaken,
-      },
-    });
-  };
-
-  const applyContentModerationStatus = async () => {
-    const mappedStatus = actionToModerationStatus[actionTaken];
-
-    try {
-      await applyAiModerationDecisionService(
-        reportContentId,
-        targetType,
-        mappedStatus,
-      );
-    } catch (error) {
-      console.error(
-        "[adminModerateReport] Failed to apply content moderation status",
-        {
-          decisionId,
-          reportId,
-          reportContentId,
-          targetType,
-          mappedStatus,
-          error,
-        },
-      );
-    }
-  };
-
-  try {
-    switch (actionTaken) {
-      case "BAN_TEMP": {
-        const expiresAt = new Date(Date.now() + (banDurationMs as number));
-
-        await prisma.$transaction(async (tx) => {
-          const existingPermBan = await tx.ban.findFirst({
-            where: { userId: reportTargetUserId, banType: "PERM" },
-          });
-
-          if (existingPermBan)
-            throw new HttpError("User already has a permanent ban", 409);
-
-          const existingTempBan = await tx.ban.findFirst({
-            where: {
-              userId: reportTargetUserId,
-              banType: "TEMP",
-              expiresAt: { gt: new Date() },
-            },
-          });
-
-          if (existingTempBan)
-            throw new HttpError("User already has an active temp ban", 409);
-
-          await tx.ban.create({
-            data: {
-              userId: reportTargetUserId,
-              title,
-              reasons,
-              banType: "TEMP",
-              bannedBy: "ADMIN_MODERATION",
-              expiresAt,
-              durationMs: banDurationMs,
-            },
-          });
-
-          await tx.user.update({
-            where: { id: reportTargetUserId },
-            data: { status: "SUSPENDED" },
-          });
-        });
-
-        const meta = {
-          reportId,
-          targetContentId: reportContentId,
-          targetContentType: targetType,
-          action: "BAN_TEMP",
-          title,
-          reasons,
-          expiresAt,
-          durationMs: banDurationMs,
-        };
-
-        await updateReportStatus("RESOLVED", "BAN_TEMP", meta);
-        await applyContentModerationStatus();
-        await queueDeleteContentIfNeeded(meta);
-
-        await moderationMetricsQueue.add(
-          "BAN_TEMP",
-          { userId: reportTargetUserId },
-          {
-            removeOnComplete: true,
-            removeOnFail: false,
-            jobId: makeJobId("moderationMetrics", decisionId, "BAN_TEMP"),
-          },
-        );
-
-        await routeNotification({
-          recipientId: reportTargetUserId,
-          actorId: reviewedBy,
-          event: "STRIKE",
-          target: {
-            entityType: "USER",
-            entityId: reportTargetUserId,
-          },
-          meta,
-        });
-
-        await publishSocketDisconnect(reportTargetUserId);
-
-        break;
-      }
-
-      case "BAN_PERM": {
-        await prisma.$transaction(async (tx) => {
-          await tx.ban.create({
-            data: {
-              userId: reportTargetUserId,
-              title,
-              reasons,
-              banType: "PERM",
-              bannedBy: "ADMIN_MODERATION",
-            },
-          });
-
-          await tx.user.update({
-            where: { id: reportTargetUserId },
-            data: { status: "TERMINATED" },
-          });
-        });
-
-        const meta = {
-          reportId,
-          targetContentId: reportContentId,
-          targetContentType: targetType,
-          action: "BAN_PERM",
-          title,
-          reasons,
-        };
-
-        await updateReportStatus("RESOLVED", "BAN_PERM", meta);
-        await applyContentModerationStatus();
-        await queueDeleteContentIfNeeded(meta);
-
-        await moderationMetricsQueue.add(
-          "BAN_PERM",
-          { userId: reportTargetUserId },
-          {
-            removeOnComplete: true,
-            removeOnFail: false,
-            jobId: makeJobId("moderationMetrics", decisionId, "BAN_PERM"),
-          },
-        );
-
-        await routeNotification({
-          recipientId: reportTargetUserId,
-          actorId: reviewedBy,
-          event: "STRIKE",
-          target: {
-            entityType: "USER",
-            entityId: reportTargetUserId,
-          },
-          meta,
-        });
-
-        await publishSocketDisconnect(reportTargetUserId);
-
-        break;
-      }
-
-      case "WARN": {
-        const expiresAt = new Date(Date.now() + (warningDurationMs as number));
-
-        await prisma.warning.create({
-          data: {
-            userId: reportTargetUserId,
-            title,
-            reasons,
-            warnedBy: "ADMIN_MODERATION",
-            expiresAt,
-          },
-        });
-
-        const meta = {
-          reportId,
-          targetContentId: reportContentId,
-          targetContentType: targetType,
-          title,
-          reasons,
-          expiresAt,
-        };
-
-        await updateReportStatus("RESOLVED", "WARN", meta);
-        await applyContentModerationStatus();
-        await queueDeleteContentIfNeeded(meta);
-
-        await moderationMetricsQueue.add(
-          "WARN",
-          { userId: reportTargetUserId },
-          {
-            removeOnComplete: true,
-            removeOnFail: false,
-            jobId: makeJobId("moderationMetrics", decisionId, "WARN"),
-          },
-        );
-
-        await routeNotification({
-          recipientId: reportTargetUserId,
-          actorId: reviewedBy,
-          event: "WARN",
-          target: {
-            entityType: "USER",
-            entityId: reportTargetUserId,
-          },
-          meta,
-        });
-
-        break;
-      }
-
-      case "IGNORE": {
-        const meta = { title, reasons };
-
-        await updateReportStatus("DISMISSED", "IGNORE", meta);
-        await applyContentModerationStatus();
-
-        await moderationMetricsQueue.add(
-          "IGNORE",
-          { userId: reportTargetUserId },
-          {
-            removeOnComplete: true,
-            removeOnFail: false,
-            jobId: makeJobId("moderationMetrics", decisionId, "IGNORE"),
-          },
-        );
-
-        break;
-      }
-    }
-
-    await clearReportsCache();
-  } catch (error) {
-    await Report.findOneAndUpdate(
-      {
-        _id: foundReport._id,
-        status: "PENDING",
-        reviewedBy,
-      },
-      {
-        reviewedBy: null,
-        reviewedAt: null,
-        $unset: { reviewComment: 1 },
-      },
-    );
-
-    throw error;
-  }
-};
-
-export default adminModerateReport;
